Message: improve adding attachment
`message::attach` takes an `istream` and reads the content into a
stringstream. This unfortunately only reads till the first whitespace
character.

This adds a new method to pass the content directly as a `string`.
It should get discussed if the old method should be deprecated.

Also fix the example to read the images as binary and pass them
to the new function.