provide the smtp servers text on accepting mail
When an SMTP server accapts a mail message it responds with a positive
completion status (2xy code) and some kind of informative text, usually
containing a queued ID, e.g. "Ok: queued as 9F04854013C".

One might want to keep this text or save it to a log file as a proof
that the mail message has been sent correctly.

In order to do so, this patch adds a return string to the smpt::submit()
method, which returns exactly this text.