use `unsigned char` for char classification
The C standard does not define how `isalpha()` behaved when we pass it a negative number. It deals with ASCII only. GLIBC tries to handle this by testing it as the current locale, which is... not something the standard demands.

MSVC is more strict - it just throws.

So - all these functions need to have a `uch` value - ugly, and simple solution.

Some RTFM:
https://news.ycombinator.com/item?id=28703525
https://drewdevault.com/2020/09/25/A-story-of-two-libcs.html